Update CMake configuration, enhance logging, and improve packet formatting
- Set default C++ version to 23 in CMakeLists.txt for better compatibility.
- Adjusted fmt library usage to ensure proper formatting of byte data in packet.hpp and rtpmidipacket.hpp.
- Updated logger.hpp to include alternative print functionality for better compatibility.
- Added libsqlite3-dev to build dependencies in Dockerfiles for Debian and Ubuntu.
- Updated changelog for versioning consistency.

+++ include/rtpmidid/rtpmidipacket.hpp
@@ -110,7 +110,7 @@ public:
     if (!is_midi_packet()) {
       std::string first_12_bytes_hex;
       for (int i = 0; i < 12; i++) {
-        first_12_bytes_hex += FMT::format("{:02x} ", data[i]);
+        first_12_bytes_hex += FMT::format("{:02x} ", unsigned(data[i]));
       }
       return FMT::format("RTP Packet: Invalid MIDI packet {}",
                          first_12_bytes_hex);
@@ -207,7 +207,7 @@ public:
     if (!is_command_packet()) {
       std::string first_12_bytes_hex;
       for (int i = 0; i < 12; i++) {
-        first_12_bytes_hex += FMT::format("{:02x} ", data[i]);
+        first_12_bytes_hex += FMT::format("{:02x} ", unsigned(data[i]));
       }
       return FMT::format("RTP Command Packet: Invalid Command packet {}",
                          first_12_bytes_hex);
@@ -274,7 +274,7 @@ public:
     if (!is_command_packet()) {
       std::string first_12_bytes_hex;
       for (int i = 0; i < 16; i++) {
-        first_12_bytes_hex += FMT::format("{:02x} ", data[i]);
+        first_12_bytes_hex += FMT::format("{:02x} ", unsigned(data[i]));
       }
       return FMT::format("RTP Command Packet: Invalid Command packet {}",
                          first_12_bytes_hex);
@@ -354,7 +354,7 @@ public:
     if (!is_command_packet()) {
       std::string first_12_bytes_hex;
       for (int i = 0; i < 12; i++) {
-        first_12_bytes_hex += FMT::format("{:02x} ", data[i]);
+        first_12_bytes_hex += FMT::format("{:02x} ", unsigned(data[i]));
       }
       return FMT::format("RTP Command Packet: Invalid Command packet {}",
                          first_12_bytes_hex);
